requirements.txt
python_telegram_bot
redis
flask
python-dotenv
pillow
requests